fixbug: tmp451 temperature wrong when under zero

Former-commit-id: bb2192695d77d99002a0874d73ef47da0136bef3

+++ SA5AIUC10/Src/tmp451.h
@@ -9,6 +9,6 @@
 #define __TMP451_H__
 
 extern void tmp451_init(void);
-extern void tmp451_set_temp(uint8_t soc, uint8_t board);
+extern void tmp451_set_temp(int soc, int board);
 
 #endif /* DS1307_H_ */
